# swagger/api.yml
swagger: '2.0'
info:
  title: DynamicPush MEAN API
  description: DynamicPush API documentation for Node Express
  version: 0.1.0
host: localhost
schemes:
  - http
basePath: /api
produces:
  - application/json
paths:
...
  /applications/{applicationId}/segments:
    get:
      summary: list segments
      description: |
        Returns list of segments inside application
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
      tags:
        - Segments
      responses:
        '200':
          description: An array of segmetns
          schema:
            type: array
            items:
              $ref: '#/definitions/Segment'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: create segment
      description: |
        Creates a segment with basic properties
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
        - name: body
          description: json object for segment
          required: true
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
                description: name of the segment
                example: Age Group
              filter:
                type: object
      tags:
        - Segments
      responses:
        '200':
          description: An object of segment
          schema:
            $ref: '#/definitions/Segment'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
  /applications/{applicationId}/segments/{segmentId}:
    get:
      summary: get a segment detail
      description: |
        Get a segment detail
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
        - name: segmentId
          required: true
          in: path
          type: string
      tags:
        - Segments
      responses:
        '200':
          description: An object of segment
          schema:
            $ref: '#/definitions/Segment'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: delete a segment
      description: |
        delete a segment
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
        - name: segmentId
          required: true
          in: path
          type: string
      tags:
        - Segments
      responses:
        '200':
          description: An object of segment
          schema:
            $ref: '#/definitions/Segment'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: update segment
      description: |
        Update a segment
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
        - name: segmentId
          required: true
          in: path
          type: string
        - name: body
          description: json object for segment
          required: true
          in: body
          schema:
            $ref: '#/definitions/Segment'
      tags:
        - Segments
      responses:
        '200':
          description: An object of segment
          schema:
            $ref: '#/definitions/Segment'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
  /applications/{applicationId}/filters:
    get:
      summary: list filters
      description: |
        Returns list of filters inside application
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
      tags:
        - Filters
      responses:
        '200':
          description: An array of filters
          schema:
            type: array
            items:
              $ref: '#/definitions/Filter'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: create filter
      description: |
        Creates a filter with basic properties
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
        - name: body
          description: json object for filter
          required: true
          in: body
          schema:
            type: object
            properties:
              body:
                type: object
                description: object for filter similar to mongoose query format
      tags:
        - Filters
      responses:
        '200':
          description: An object of filter
          schema:
            $ref: '#/definitions/Filter'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
  /applications/{applicationId}/filters/{filterId}:
    get:
      summary: get a filter detail
      description: |
        Get a filter detail
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
        - name: filterId
          required: true
          in: path
          type: string
      tags:
        - Filters
      responses:
        '200':
          description: An object of filter
          schema:
            $ref: '#/definitions/Filter'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: delete a filter
      description: |
        delete a filter
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
        - name: filterId
          required: true
          in: path
          type: string
      tags:
        - Filters
      responses:
        '200':
          description: An object of filter
          schema:
            $ref: '#/definitions/Filter'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: update filter
      description: |
        Update a filter
      parameters:
        - name: applicationId
          required: true
          in: path
          type: string
        - name: filterId
          required: true
          in: path
          type: string
        - name: body
          description: json object for filter
          required: true
          in: body
          schema:
            $ref: '#/definitions/Filter'
      tags:
        - Filters
      responses:
        '200':
          description: An object of filter
          schema:
            $ref: '#/definitions/Filter'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Application:
    type: object
    properties:
      image:
        type: string
        description: icon url
      appName:
        type: string
        description: Name of the application
      packageName:
        type: string
        description: Mobile app package name
      apiKey:
        type: string
        description: Key used for mobile api
      apiSecret:
        type: string
        description: Secret used for mobile api
      fcmServerKey:
        type: string
        description: FCM Server Key used for push notifications on Android
      environment:
        type: string
        description: development or production
...
  Segment:
    type: object
    properties:
      _id:
        type: string
      name:
        type: string
        example: Age Group
      filter:
        type: object
      application:
        $ref: '#/definitions/Application'
  Filter:
    type: object
    properties:
      _id:
        type: string
      body:
        type: object
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string